requirements.txt
pytest==6.1.2
selenium==3.141
requests==2.24
simple-settings==1.0.0
PyHamcrest==2.0.2
pytest-xdist==2.1.0
pytelegrambotapi==3.7.3
pytz==2020.4
